requirements.txt
numpy==1.23.3
opencv-contrib-python==4.6.0.66
Pillow==9.2.0

